Disse selectorene hører mer hjemme under ducks/flyt
