+++ src/ducks/flyt/selectors.test.js
@@ -1,6 +1,8 @@
 import * as selectors from './selectors';
 import MKV from '../../melosyskodeverk';
 
+import * as DucksTestUtils from '../test-utils';
+
 describe('FlytSelectors', () => {
   describe('UtpekingVurderingSelector', () => {
     const lagState = behandlingstema => ({
@@ -35,6 +37,33 @@ describe('FlytSelectors', () => {
         const state = lagState(behandlingstema);
         expect(selectors.UtpekingVurderingSelector(state)).toBe(forventet);
       });
+    });
+  });
+
+  describe('ErIArtikkel13_1FlytSelector', () => {
+    each([
+      [
+        true,
+        MKV.Koder.behandlinger.behandlingstema.ARBEID_FLERE_LAND,
+      ],
+      [
+        false,
+        MKV.Koder.behandlinger.behandlingstema.UTSENDT_SELVSTENDIG,
+      ],
+    ]).it('returnerer korrekt verdi', (forventetResultat, behandlingstema) => {
+      const state = DucksTestUtils.lagState({
+        behandlinger: {
+          data: {
+            oppsummering: {
+              behandlingstema: {
+                kode: behandlingstema,
+              },
+            },
+          },
+        },
+      });
+
+      expect(selectors.ErIArtikkel13_1FlytSelector(state)).toEqual(forventetResultat);
     });
   });
 });
